+++ docs/manual/intro-to-pipeline.md
@@ -51,15 +51,18 @@ $ pipcook run /path/to/your/pipeline-config.json
 The trained model will generate an `output` directory under [`cwd(3)`](https://linux.die.net/man/3/cwd):
 
 ```
-📂output
-   ┣ 📂logs
-   ┣ 📂model
-   ┣ 📜package.json
-   ┣ 📜metadata.json
-   ┗ 📜index.js
+output
+├─ logs
+├─ nodejs
+├─ package.json
+└─ wasm
 ```
 
-To get started with your trained model, follow the below steps:
+There are two entry-points here, one for nodejs powered by boa; another one is used for wasm (WebAssembly), a universial virtual machine, powered by TVM.
+
+**Note:** TVM does not have linux wheel at this time, therefore, we will only export WASM format for mac platform at this time. We will provide a detailed guide for linux users to build TVM from source, if you want to try it out under linux.
+
+To get started with your trained model, follow the below steps under `nodejs` and `wasm` seperately:
 
 ```sh
 $ npm install
@@ -77,3 +80,5 @@ Once the output is initialized, just `import` it as the following:
 import * as predict from './output';
 predict('your input data');
 ```
+
+**Note:** The wasm format is under experiment, in theory, the generated model should be possible to run with GPU via webGPU. However, the standard for webGPU is not finalized. We only enable cpu mode by default at this time.
